Feature/17 fast bid homepage

+++ database/migrations/2018_12_05_105804_create_is_auction_closed_function.php
@@ -12,7 +12,7 @@ class CreateIsAuctionClosedFunction extends Migration
     public function up()
     {
         statement('
-            CREATE FUNCTION dbo.is_auction_closed (@item_id INT) RETURNS BIT
+            CREATE FUNCTION dbo.is_auction_closed (@item_id BIGINT) RETURNS BIT
                 AS BEGIN RETURN (
                     SELECT auction_closed FROM items where id = @item_id
                 )
